UBUNTU: SAUCE: platform: tegra: Add support for mask with 1s
BugLink: https://bugs.launchpad.net/bugs/2028400

Currently, the value of prod setting are applied if mask
bit is 0 for given value. This is non-traditional style of
defining mask.

Add support a traditional way of defining mask like if mask bits
are 1s then only apply the value of those bits.

	reg = (reg & ~mask) | (reg & mask);

This mechanism is selected via DT property to have backward
compatibility. The DT property is "mask-one-style" which can be
added under prod-setting node.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Signed-off-by: Abhilash G <abhilashg@nvidia.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Ian May <ian.may@canonical.com>
Signed-off-by: Jacob Martin <jacob.martin@canonical.com>