BACKPORT: KVM: arm64: Document MTE capability and ioctl
A new capability (KVM_CAP_ARM_MTE) identifies that the kernel supports
granting a guest access to the tags, and provides a mechanism for the
VMM to enable it.

A new ioctl (KVM_ARM_MTE_COPY_TAGS) provides a simple way for a VMM to
access the tags of a guest without having to maintain a PROT_MTE mapping
in userspace. The above capability gates access to the ioctl.

Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Steven Price <steven.price@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20210621111716.37157-7-steven.price@arm.com
(cherry picked from commit 04c02c201d7e8149ae336ead69fb64e4e6f94bc9)
[willdeacon@: Dropped missing context]
Signed-off-by: Will Deacon <willdeacon@google.com>
Bug: 192636784
Change-Id: I489a4cc12cad2b4664b131d4d65c9d12024eaa48